+++ bht/catalog_tools.py
@@ -187,8 +187,11 @@ def rows_to_catstring(events_list, catalog_name, columns=None):
         v = hpevent_parameters[c]
         r_string += f'# Parameter {i}: id:column{i}; name: {v["col_name"]}; size:{v["size"]}; type:{v["type"]};\n'
 
+    if len(events_list) == 0:
+        return r_string
+
     # Rewrite every dict in list
-    events_list = [dict_to_dict(e) for e in events_list]
+    events_list = [dict_to_dict(e, columns) for e in events_list]
 
     # store max lengths in a dictionnary with same keys as events
     values_lengths = {k: [] for k in events_list[0].keys()}